requirements.txt
streamlit
gitpython